+++ Backend/Tabibi.API/Controllers/AdminController.cs
@@ -37,7 +37,7 @@ namespace Tabibi.API.Controllers
         }
 
         [HttpGet("doctors/{doctorId}")]
-        public async Task<IActionResult> GetDoctorDetail(int doctorId)
+        public async Task<IActionResult> GetDoctorDetail(long doctorId)
         {
             var doctor = await adminService.GetDoctorDetail(doctorId);
             if (doctor == null) return NotFound();
@@ -45,14 +45,14 @@ namespace Tabibi.API.Controllers
         }
 
         [HttpGet("doctors/{doctorId}/changes")]
-        public async Task<IActionResult> GetDoctorChanges(int doctorId)
+        public async Task<IActionResult> GetDoctorChanges(long doctorId)
         {
             var changes = await adminService.GetDoctorChanges(doctorId);
             return Ok(changes);
         }
 
         [HttpPatch("doctors/{doctorId}/verify")]
-        public async Task<IActionResult> VerifyDoctor(int doctorId, [FromBody] ReviewDoctorRequestDTO request)
+        public async Task<IActionResult> VerifyDoctor(long doctorId, [FromBody] ReviewDoctorRequestDTO request)
         {
             var res = await adminService.VerifyDoctor(doctorId, request);
             if (!res.IsSuccess)
