+++ Backend/Tabibi.Application/DTOs/AppointmentFilterDTO.cs
@@ -14,8 +14,8 @@ public class AppointmentFilterDTO
 
 public class AppointmentListDTO
 {
-    public int DoctorId { get; set; }
-    public int AppointmentId { get; set; }
+    public long DoctorId { get; set; }
+    public long AppointmentId { get; set; }
     public string DoctorName { get; set; } = string.Empty;
     public string PatientName { get; set; } = string.Empty;
     public DateTime ScheduledAt { get; set; }
@@ -26,7 +26,7 @@ public class AppointmentListDTO
     public string? Notes { get; set; }
     public string? DoctorProfilePictureUrl { get; set; }
     public string? PatientProfilePictureUrl { get; set; }
-    public int? SessionId { get; set; }
+    public long? SessionId { get; set; }
     public int? ReviewRating { get; set; }
     public string? ReviewComment { get; set; }
 }
