+++ Backend/Tabibi.Application/Interfaces/IChatService.cs
@@ -6,12 +6,12 @@ namespace Tabibi.Application.Interfaces;
 
 public interface IChatService
 {
-    Task<ChatAccessResult> ValidateAccess(int sessionId, string userId);
-    Task<List<ChatMessageDTO>> GetHistory(int sessionId);
-    Task<ChatSessionDetailsDTO?> GetSessionDetails(int sessionId);
+    Task<ChatAccessResult> ValidateAccess(long sessionId, string userId);
+    Task<List<ChatMessageDTO>> GetHistory(long sessionId);
+    Task<ChatSessionDetailsDTO?> GetSessionDetails(long sessionId);
     Task<List<ChatSessionSummaryDTO>> GetUserSessions(string userId, string role);
-    Task<ChatMessage> SaveMessage(int sessionId, string role, string content, bool isSystemMessage = false);
-    Task<ChatSession> StartOrGetSessionAsync(string patientUserId, int doctorId, bool isCompanyPaid = false);
-    Task<ChatSession> StartOrGetAISessionAsync(string patientUserId, int? sessionId = null);
-    Task<ChatSession> FollowUpSessionAsync(int sessionId, string patientUserId);
+    Task<ChatMessage> SaveMessage(long sessionId, string role, string content, bool isSystemMessage = false);
+    Task<ChatSession> StartOrGetSessionAsync(string patientUserId, long doctorId, bool isCompanyPaid = false);
+    Task<ChatSession> StartOrGetAISessionAsync(string patientUserId, long? sessionId = null);
+    Task<ChatSession> FollowUpSessionAsync(long sessionId, string patientUserId);
 }
