+++ Backend/Tabibi.Application/Interfaces/IPublicService.cs
@@ -6,5 +6,5 @@ public interface IPublicService
 {
     Task<List<SpecialtyDTO>> GetSpecialtiesAsync();
     Task<PaginatedResultDTO<DoctorListDTO>> GetDoctorsAsync(DoctorSearchFilterDTO filter);
-    Task<DoctorListDTO?> GetDoctorByIdAsync(int doctorId);
+    Task<DoctorListDTO?> GetDoctorByIdAsync(long doctorId);
 }
