+++ Backend/Tabibi.Application/Interfaces/IReviewService.cs
@@ -4,6 +4,6 @@ namespace Tabibi.Application.Interfaces;
 
 public interface IReviewService
 {
-    Task<ServiceResult<PagedReviewsDTO>> GetDoctorReviewsAsync(int doctorId, int page = 1, int pageSize = 10);
+    Task<ServiceResult<PagedReviewsDTO>> GetDoctorReviewsAsync(long doctorId, int page = 1, int pageSize = 10);
     Task<ServiceResult> SubmitReviewAsync(string userId, CreateReviewDTO dto);
 }
